bpf: common: move structs & definitions into feature-specific headers
Maintain the keys/values along with their respective maps in
feature-specific header files. This also enables better CODEOWNERS
granularity.

This just tackles the low-hanging fruit, some others are more involved
or worth their own patch.
