+++ bpf/lib/edt.h
@@ -8,6 +8,27 @@
 #include "common.h"
 #include "time.h"
 
+#define DIRECTION_EGRESS 0
+#define DIRECTION_INGRESS 1
+
+struct edt_id {
+	__u32		id;
+	__u8		direction;
+	__u8		pad[3];
+};
+
+struct edt_info {
+	__u64		bps;
+	__u64		t_last;
+	union {
+		__u64	t_horizon_drop;
+		__u64	tokens;
+	};
+	__u32		prio;
+	__u32		pad_32;
+	__u64		pad[3];
+};
+
 struct {
 	__uint(type, BPF_MAP_TYPE_HASH);
 	__type(key, struct edt_id);
