+++ bpf/lib/ipsec.h
@@ -18,6 +18,11 @@
 /* We cap key index at 4 bits because mark value is used to map ctx to key */
 #define MAX_KEY_INDEX 15
 
+/* encrypt_config is the current encryption context on the node */
+struct encrypt_config {
+	__u8 encrypt_key;
+} __packed;
+
 struct {
 	__uint(type, BPF_MAP_TYPE_ARRAY);
 	__type(key, __u32);
